services/enhance/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pillow==10.3.0
pydantic-settings==2.2.1
python-multipart==0.0.9
pyyaml==6.0.1
respx==0.21.1
pytest==8.2.2
pytest-asyncio==0.23.7
